+++ serialization/src/main/kotlin/net/corda/serialization/internal/DefaultWhitelist.kt
@@ -5,7 +5,6 @@ import net.corda.core.utilities.NetworkHostAndPort
 import org.apache.activemq.artemis.api.core.SimpleString
 import rx.Notification
 import rx.exceptions.OnErrorNotImplementedException
-import sun.security.x509.X509CertImpl
 import java.security.cert.CRLReason
 import java.util.*
 
@@ -62,7 +61,6 @@ object DefaultWhitelist : SerializationWhitelist {
                     StackTraceElement::class.java,
 
                     // Implementation of X509Certificate.
-                    X509CertImpl::class.java,
                     CRLReason::class.java
             )
 }