+++ webserver/src/main/java/CordaWebserverCaplet.java
@@ -9,6 +9,7 @@ import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
+import java.util.stream.Stream;
 
 public class CordaWebserverCaplet extends Capsule {
 
@@ -174,8 +175,8 @@ public class CordaWebserverCaplet extends Capsule {
 
     private static void checkJavaVersion() {
         String version = System.getProperty("java.version");
-        if (version == null || !version.startsWith("1.8")) {
-            System.err.printf("Error: Unsupported Java version %s; currently only version 1.8 is supported.\n", version);
+        if (version == null || Stream.of("1.8", "11").noneMatch(version::startsWith)) {
+            System.err.printf("Error: Unsupported Java version %s; currently only version 1.8 or 11 is supported.\n", version);
             System.exit(1);
         }
     }